Move consent tests to dedicated test_consent_views.py
Reduces test_chatbot_views.py back to its pre-PR line count by
extracting TestConsentWarnings and test_chatbot_list_shows_consent_status_column
into a focused module, addressing CodeScene's Lines of Code and Low
Cohesion findings.
